feat: do not commit most generated files
Inspired by @harrysolovay, as much as possible, generated files should not be committed.

This change allows specify at the `FileBase` level if a synthesized file should be committed or not. By default, all files are automatically added to .gitignore so they
will not be committed.

We still need to figure out how to bootstrap when we don't have a package.json, but this is the same as bootstrapping a new project, so that's good.

Also, make all constructs always connect to the `Project` so basically we now have a flat tree.