fix(node): operation not permitted: unlink 'node_modules/projen'
After yarn installs `node_modules/projen`, it is no longer a file but a directory
so unlinking it fails.

Now, the heuristics are as follows:

If `node_modules/projen/package.json` does not exist or `node_modules/projen` is not a directory (could be a symlink from another run), we recreate it as a symlink to the executing cli.

Then, after synthesis, before running `yarn install`, if there's a symlink, we remove it.
We also run `yarn install` with `--check-files` to verify the integrity of `node_modules`.

Additionally, in CI builds, we run yarn install with `--frozen-lockfile`.

Fixes #93