fix(node): wrong npmIgnore field used (#181)
*Issue #, if available:* Closes #139 

*Description of changes:* Fixed npmIgnore to properly remove the .npmignore file if set to false.

It might be worth changing the name of one of these fields (which differ just by capitalization) to eliminate this kind of confusion. Perhaps change `npmignore` to `npmIgnorePatterns`? It's a little more verbose but not by much.

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.