feat: typescript nextjs projects (#197)
*Issue #, if available:* N/A

*Description of changes:* Added a nextjs-ts project option which extends the typescript-app project class. Grouped some options together between the two nextjs project types, and implemented basic Next.js tsconfig project defaults so that both nextjs and nextjs-ts project types work out of the box.

----

Some notes:

I made a slight modification of TypeScriptConfig so that I can mark `tsconfig.json` as mutable, since the 'next build' command fails if it's immutable. The reason for this I believe is that the 'next' command tries to automatically change your tsconfig file if it sees that some options set incorrectly based on what Next.js's build system wants. I could send a feature request to the Next.js team to see if they can enable a CLI flag that skips this behavior, but for now this workaround makes the most sense.

There is a two-line file named `next-env.d.ts` that must be added to the root directory for TypeScript Next.js projects to successfully build. I included this in the PR, but it's worth suggesting that we should also eventually add an escape hatch so a user can omit or change the `next-env.d.ts` file if needed -- but I'm not sure what that mechanism should be, so I've left that unimplemented for now.

~Lastly, many Next.js projects don't have test directories from what I've seen, so I didn't include sample code for any. But this causes eslint to complain, and changing how the default eslint command behaves warrants a separate PR (see #182), and using `eslint: false` can lead to #180, so I ended up adding a small piece of cruft in the project constructor for now:~

```
this.addTestCommand('echo 0');
```

(I ended up overriding the default TypeScriptProject build command so this isn't an issue)

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.