fix: start menu does not respect package manager (#228)
Add a `command` to each start menu entry in `package.json` and use it instead of assuming this is a yarn script. This opens the door for decoupling start menu from yarn/npm and allowing it to be used in other project types as an entry point.

This change also adds `ScriptOptions` to `node.addScript()` such that it is possible
to register a start menu entry together with adding scripts to the project.

Fixes #217

BREAKING CHANGE: `node.addScript(script, command)` now takes a single command instead of variable arguments. Use `node.addScript(script, [s1,s2].join(' && '))` to build a command from multiple statements.