feat(node): deprecate "xxxDependencies" in favor of "xxxDeps" (warning)
Mark the `xxxDependencies` options as deprecated and issue a warning during synthesis. The alternative is to use the `xxxDeps` option which allows managing versions in `package.json` file. This is a superior approach as it allows external tools such as dependabot to manage your module versions.

Instead of:

  dependencies: { 'express': Server.caret('2.0.0') }

Use:

  deps: [ 'express' ]

Or if you need a specific version (unlikely):

  deps: [ 'express@^2' ]

When a version is not specified, projen will behave like `npm i` or `yarn add`. It will install the latest version and record it as a caret dependency in your `package.json` file. From here on out,
the version will be managed in your `package.json` file.