feat: subprojects (#332)
Introduce an optional `parent` option to `Project` and `outdir` to allow implementing project trees.

The `CompositeProject` approach which was introduced in #289 has a limitation in which projects could not query who is their parent (because potentially a child can be added at any time). Also, it technically allowed the same child project to be added to multiple parents. Therefore, these classes are no longer available.

This change also moves `outdir` from the various `synth()` methods to `ProjectOptions`.

This early-binding approach is also more in-line with how CDK code is written, whereas during initialization it is possible to reason about the full synthesis context.

Additional APIs introduced in this change are `project.files` and `project.findFile(path)` which can be used to inspect the project's contents. `findFile()` can find files in projects and subprojects.

BREAKING CHANGE: The `CompositeProject` and `ProjectComponent` classes have been superseded by subprojects in order to allow safely accessing the parent project during subproject initialization phase. Use `new Project({ parent, outdir })` to define this relationship.
* The various `synth()` methods no longer accept an `outdir`. Instead, specify `outdir` in the project options.