feat(node): rebuild bot (#349)
Installs a GitHub workflow which is triggered by adding a comment on a pull request which reads `@projen rebuild`. The workflow will check out the PR branch and build it (with no anti-tamper checks), and then it will push the changes back to the PR branch.

This is particularly useful in order to update test snapshots and other generated files such as `API.md`.

BREAKING CHANGE: `bootstrapTask` is no longer available. To obtain the bootstrapping steps for GitHub workflows, use `installWorkflowSteps`.
* **node:** the `workflowBootstrapSteps` is no longer available. The equivalent is `[ { uses: 'actions/checkout@v2' }, project.installWorkflowSteps ]`.
* **node:** the `NodeBuildWorkflowOptions` struct is no now internal.