chore: integration tests
Integration tests are a complete snapshot of the synthesis of real projects.

Ideally most of the coverage is achieved through unit tests, but this is a cheap way to capture unw

To add sample real projects, just add files with the `xxx.projenrc.js` extension under `test/integration`. If the file is in a subdirectory, the entire directory will be staged.

The `xxx.projenrc.js` file will be renamed `.projenrc.js` and the `projen` CLI will be executed. The resulting deep directory snapshot will be taken as a jest snapshot.