feat: Make IgnoreFile stateful to allow for overrides (#435)
I faced an issue that did not allow me to override the gitignore file so that it would ignore the projen generated files. This is due to how the IgnoreFile currently works. With the modification I made, we can allow users to override the IgnoreFile anywhere down the road. 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.