feat(cli): allow specifying the location of `.projenrc.js` using `--rc` (#440)
It is now possible to set the location of the `.projenrc.js` file using the `--rc` switch in the CLI. This can be used to synthesize project directories without the rc file in them (e.g. for testing).

This requires relaxing the constraint that `outdir` of a project may only be `.` if `.projenrc.js` exists in the same directory (since this is the default).

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.