feat(cli): improved bootstrapping
Looking towards supporting multiple languages for projenrc, we need to give more control to the project classes (over the cli) when it comes to bootstrapping new projects.

This changes how `new` is implemented such that synthesis is performed by calling `.synth()` on a project object, and not by using the CLI's synth code path, which is tailored for the javascript version of .projenrc.

For example, it will now be possible to implement a `JavaProject` which will synthesize a java directory structure and a manifest that will take a java dependency on "projen" from maven. It will also be configured to read the projenrc from a java file within the source tree.

This solves the bootstrapping problem for compiled languages: if projenrc need to be compiled in order to be synthesized, we will need the full toolchain and its configuration in order to compile and execute it. If we want to be able to use any (supported) language we want for projenrc, we need to first synthesize the toolchain environment, and that's exactly what projects are doing.