feat: publish projen to maven and pypi (#466)
Start publishing the projen library to Maven Central (Java) and PyPI (Python).

The option names in `JsiiProjectOptions` for publishing to package managers were renamed to:
- `java` => `publishToMaven`
- `python` => `publishToPypi`
- `dotnet` => `publishToNuget`

The rename has two reasons: (1) the new names are grouped together under `publishTo`, and are more accurate since they represent the target package manager and not the language; and (2) when jsii tries to compile the generated java code, using name `java` for a field in a struct yields compilation errors (https://github.com/aws/jsii/issues/2405).

The `python` and `dotnet` options have been `@deprecated`, so they should still work, but due to the above mentioned bug, the `java` option have been completely removed. To help users migrate, if they use `java`, an error is thrown with the new name.

BREAKING CHANGE: the `java` option in jsii projects is now called `publishToMaven`.