fix: output to STDOUT crashes jsii languages
See https://github.com/aws/jsii/issues/2406 for details.

Until this is resolved in jsii, we bluntly hook `process.stdout.write()` and redirect the stream to STDERR. We do the same thing when we `exec()` shell commands.

NOTE: the task runtime still inherits STDOUT, so this applies only to logs coming from libraries.