feat(java): support projenrc in java
If `projenrc` is set to `true` for java projects, a `synth` task will be added which will execute the Java program under `src/test/java/projenrc.java`. This enables writing your projenrc files in java (for java projects).

Enabling this will add a test dependency on `com.github.eladb/projen` and a task called `synth` which will compile the code and run it.

So your workflow is to update `projenrc.java` and run `pj synth`.

In a subsequent commit, we will auto-detect the `synth` task and use it when `pj` is executed without any subcommand, so projects can customize their synthesize task.