feat(java): sub-dependencies for maven plugins (#469)
In maven, it is possible to specify dependencies that only apply to a specific plugin. This commit adds that capability.

NOTE: dependencies of plugins are not recorded in the project's dependency model at this point.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.