fix: bump task fails on subprojects (#477)
When running `yarn bump` on a subproject, standard-version executes projen in the subproject path. As the .projenrc.js only exists on the root project, this fails. This patch changes to the relative path of the root project before running projen.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.