feat(java): projenrc in java
From now, java projects will always generate a projenrc file in java based on the options passed to `projen new` (and default values for mandatory arguments).

To allow non-javascript projects to be synthesized idiomatically, the CLI will execute a task named `synth` if it exists instead of the default javascript synthesis flow. In a subsequent change, all project types will always include a `synth` task and this will be the default way for projects to be synthesized.