fix(cli): `projen` does not work for projects with a “synth” task
In order to support non-JavaScript projenrc, #505 changed the behavior of running the CLI without any arguments such that if the project had a `synth` task, it will be executed instead of running `.projenrc.js`.

This caused projects that already had a `synth` command (i.e. CDK apps, in which `synth` is used to synthesize the CDK application) to break, since the CLI would have now executed that command.

The fix is to change the name of the default task to `default` instead.

Fix #507