feat(node): make "releaseWorkflow" public (#508)
fixes: #494 
* change visibility of releaseWorkflow

That are kind of **minmal** changes necessary to archive the use case from #494. As mentioned in the issue it would be cool to build a more sophisticated way like the Tasks to manipulate existing GitHubWorkflows. But that would require a lot of reconstructions I guess. But I am happy to do that. Any feedback is welcome :)