fix(node): tags may be pushed without a bump commit (#554)
When multiple commits are pushed to the main branch at once (for whatever reason), the release workflow will try to push both the new tag and the bump commit. In such a case, the tag will be pushed but the commit will fail, which means the repo is in a corrupted state (from a bump perspective).

This change fixes the broken transaction issue by pushing the tag only after the bump commit has been pushed.

The implication is that only the latest release (in the series of commits) will succeed. Ideally we'd want to cancel the release workflow (instead of failing it), but this will be done in a future change.

Fixes #553