feat(jsii): add excludeTypescript config option (#571)
Hello 👋 

Working on a CDK constructs project where I have a setup like:

- `src/construct-one/`
- `src/construct-two/`
- etc...

Thought it would be a neat idea to put the tests inside each construct, so in `src/construct-one/test` for example. But when JSII compiled the project for JS/TypeScript, it included the `test` directory. I was able to get JSII to ignore the `test` directory by adding this to my `package.json`:

```json
{
  "jsii": {
    "excludeTypescript": ["src/**/test/*.ts"]
  }
}
```

So, this PR is to add this configuration option to `projen`.

Link to JSII docs: https://github.com/aws/jsii/blob/main/gh-pages/content/user-guides/lib-author/configuration/index.md#the-jsii-section

I'm very open to suggestions (even if it is that my location of the `test` directory is a bad idea) and let me know if the PR can be improved in any way.  In addition to the unit test lines, I did test by creating a new project, updating the `.projenrc.js` and regenerating the project.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.