fix(node-project): fix projenVersion type to correctly use a string (#608)
The NodeProject class expects a Semver object in `projenVersion`, but it's treated like a string, causing errors when stringifying to JSON files. I just corrected the type, which looks like it should be just string.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.