feat(node): mutable PR builds
Introduce an experimental feature (disabled by default) called `buildWorkflowMutable` (name is likely to change) which will cause PR builds to automatically push any changes to synthesized files or snapshots to the PR branch.

This means that if a PR is submitted and requires updates to generated files, the PR branch will automatically get updated. It also implies that anti-tamper check is disabled for PR builds. They are still enabled on the release builds.

This is done as part of the PR build which means that if the build passes and changes are pushed successfully, the PR check will be successful (doing that as a result of a different workflow will not trigger the PR build).

Related to #602 and #604