feat(node): mutable pull request builds (#624)
Introduce "mutablke builds". When enabled (by default) it will cause PR builds to automatically push any changes to synthesized files or snapshots to the PR branch. This only works for branches and will fail for forks (similar behavior to the anti-tamper check).

See attached documentation.

This means that if a PR is submitted and requires updates to generated files, the PR branch will automatically get updated. It also implies that anti-tamper check is disabled for PR builds.

This is done as part of the PR build which means that if the build passes and changes are pushed successfully, the PR check will be successful (doing that as a result of a different workflow will not trigger the PR build).

Related to #602 and #604