feat(typescript): TypeScriptProject now merges tsconfig prop into jest and eslint config (#611)
This PR changes `TypeScriptProject` so that it merges its `tsconfig` option
with the jest and eslint tsconfigs. Before, tsconfig changes had to be
duplicated in both the `tsconfigs` option and `jestOptions.typescriptConfig`
and there was no access to add includes or excludes to the eslint tsconfig.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.