feat(resolve): duck-typed resolvable objects
It's nice to be able to just pass a resolvable object that
the resolve function knows how to lazily evaluate. It already
knows how to handle functions, but being able to pass an
object means we can pass all of it's context along with it's
resolve function without having to employ our venerable old
friend `bind`.