feat(file): resolvable interface
IResolvable isn't really used for type safety, only
for type narrowing. As such it's more of a developer
guide, which hopefully leaves the door open for different
use cases and refactors in the future.

Elad made a great point that this resolvable method
essentially functions like toJSON in JavaScript. Sticking
with existing language patterns when possible is a good
practice.