feat: include projen markers in synthesized files by default (#653)
Changes the default behavior for structured files generated by projen (`JsonFile`, `YamlFile`, etc.) so that including a "Generated by projen" marker is the default behavior.

This can be disabled with the `{ marker: false }` option when you are creating a file object. If it has already been created, it can be disabled with `someFile.marker = false`.

This change does not affect text files or sample files generated by projen.

The `IMarkableFile` interface has also been removed.