feat: git tag-based versioning (#634)
To avoid the need to mutate the default branch on releases (which may be blocked by branch protection), projen will now use git tags to manage project versions.

This is achieved by creating a GitHub Release as part of the "release" workflow (executed per commit, manually or on schedule). The release includes a change log generated by standard-version and the artifacts from the `dist/` directory. We also added a synchronization mechanism in the release workflow which will skip a release if there are new commits in the release branch. This protects against a race condition that can happen if multiple commits are being merged into the release branch at the same time.

BREAKING CHANGE: version numbers are now maintained through git tags instead of `version.json`. This file should be deleted alongside the `CHANGELOG.md` file. Change log entries are now included inside a GitHub Release.