fix(release): release version is always 0.0.0
Since we are now synthesizing during build, the bumped version we set in package.json during releases is reverted when build is executed.

To address this, the `RELEASE` environment variable is set in the release workflow and that causes projen to _require_ that the version in package.json is set to an actual value.