feat: projen --watch (#728)
If `--watch` (or `-w`) is specified when running `projen`, watch `projenrc.js` and re-synthesize upon change.

NOTE: This is currently only supported for projects that use the .js version of projenrc (i.e. not java).

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.