fix: enabling dependabot prevents disabling mergify (#741)
This changes how the `mergify` configuration behaves: the `GitHub`
component now exposes an optional `mergify?: Mergify` property,
which is present when the `GitHub` is created with the `mergify: true`
(this is the default) option; and `undefined` otherwise. Then, code that
adds `mergify` rules only does so if the `GitHub` has a `mergify` defined.

In order to make this work nicer, the `Mergify` component was modified
to only create its backing `YamlFile` if at least one rule is added to it.

This is an alternative take on to #714

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.