fix(awscdk-lib): incorrect version of constructs used for CDK 2.0 (#742)
CDK 2.0 pre-releases depend on constructs@10, and not constructs@3. This
selects the version of constructs to be registered based on the CDK major
version being used.

Additionally, a new `constructsVersion` option enables user-defined
versions to be used instead of the default/inferred version.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.