feat(typescript): projenrc in typescript (#744)
TypeScript projects (and derivatives) now support `--projenrc-ts` which means that their projenrc file is written in TypeScript instead of JavaScript.

To enable this change, we needed to complete the work started to enable Java projenrc files, in which the responsibility of generating the rc file have moved from the CLI to the project type. To that end, the CLI will inject a special field (`__new__`) which will include the FQN of the project type and the CLI arguments, so that it will be possible to generate the rc file. Relevant code was extracted from the CLI and reused both for JavaScript rc files and TypeScript rc files. Additionally, the `default` task is added to tell `projen` what to do when it is executed without arguments.

Resolves #671 (no `.projenrc.js` file for java projects anymore, which was misleading)

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.