feat(tasks): add removeTask method (#757)
Simple method. Unsure of the general philosophy of projen on whether or not this should throw if you try to delete something that doesn't exist. I went with my personal choice, which is to handle that. I would be happy to change it if needed.

Also, chose `remove` instead of `delete` since `remove` is already used for other methods (e.g. `removeScript`)

Addresses a missing functionality, and was mentioned in #754 as well