fix(bump): invalid version resolved for multiple tags
If there are multiple version tags on the same commit, `git describe` would return the first tag which is lexicographically first. This is not what we want, so we now use `git tag` with a `version:refname` sort order in order to query the repo for the latest version (for real).

Fixes #762