fix(bump): invalid version selected for multiple tags (#764)
If there are multiple version tags on the same commit, `git describe` would return the first tag which is lexicographically first. This is not what we want, so we now use `git tag` with a `version:refname` sort order in order to query the repo for the latest version (for real).

Fixes #762

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.