fix(node): dependabot builds always fail (#774)
Add a condition that will only update status check if the build actually resulted in changes to the repo. This condition is added both for the "commit & push" step and the "update status" step. This means that we no longer
need the condition that only updates status checks if the PR is from a non-fork because if the push would have failed anyway and we would not get there at all.

Fixes #760