feat: duck-typed resolvable objects (#637)
It's nice to be able to just pass a resolvable object that the resolve function knows how to lazily evaluate.

The resolve function already knows how to handle functions, but being able to pass an object means we can pass all of its context along with the resolve function without having to employ our venerable old friend `bind`.

I'm unsure of how to accomplish this in a strictly-typed manner without things becoming overly complicated, or overhauling the resolve function. I don't think either of those is desirable at this point.

I've also added an `IResolvable` interface, which at this point is only used in a loose type guard.

Closes #636.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.