fix(license): apache license should not be modified (#832)
Do not modify the appendix in Apache 2.0 license. To allow other licenses to be modified (like MIT), change the substitution string to `$copyright_period` and `$copyright_owner`. If the text includes these tokens and the relevant options are not provided, fail.

Fixes #830

BREAKING CHANGE: The initializer of the `License` now takes `spdx` through the options struct instead of a positional argument. Additionally, an error is thrown if `MIT` license is selected but the `copyrightOwner` is not provided (because it is required in the license text).

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.