feat(release): support pre-release tags (#836)
Introduce a `prerelease` option which can be used bumps versions with a pre-release tag (e.g. 1.0.0-beta.1, 1.0.0-beta.2, etc).

Same as for normal versions, the latest version is resolved by querying git tags and filtering based on the pre-release identifier. If a latest version with this pre-release tag cannot be located, the `initialVersion` option is used with the pre-release tag.

Since resolving versions is now slightly more complicated, extract this logic into a "builtin" called `release/resolve-version` which is executed as part of the `bump` task. To support this, we introduce the concept of `builtins` which are executables baked into projen itself and executed as part of a task sequence.

BREAKING CHANGE: `task.toShellCommand()` has been removed, as well as the option to render npm scripts as shell commands. Maintaining this task rendition is too expensive and limits the ability to expand the task system.
* **tasks:** The `Task` initializer no longer takes `Tasks` as the first argument (no longer needed).
* **node:** The `npmTaskExecution` option has been removed.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.