fix(release): unbump is not working and fails anti-tamper
Since now we use package.json directly for bumps, using `standard-version -r 0.0.0` to revert the bump is not working any longer. Instead, add another builtin called `release/reset-version` which sets the version to `0.0.0`.

Also, move the unbump + anti-tamper check in the release workflow before the creation of the github release, so the release won't be created if anti-tamper failed.