feat(release): multiple release branches
Allow defining multiple branches from which releases are published. This is done via the following API:

    project.release.addBranch(branch, { majorVersion: 3 });

To make sure that the correct version is selected for each branch, if multiple release branches are defined, each branch must be associated with a `majorVersion`, which is used to filter the git tags when resolving the latest version.

A separate GitHub workflow is created for each release branch.

Fixes #734

BREAKING CHANGE: The `releaseBranches` has been removed in favor of `project.release.addBranch()`.
* **release:** The `initialVersion` option has been removed. The default is `v0.1.0` or `vMM.0.0`.
* **node:** The `releaseWorkflow` property is no longer available. To add jobs to the release workflow use `project.release.addJobs()`.

Misc: change projen log icon to something cutter.