feat(release): multiple release branches (#850)
Allow defining multiple branches from which releases are published. This is done via the following API:

    project.release.addBranch(branch, { majorVersion: 3 });

To make sure that the correct version is selected for each branch, if multiple release branches are defined, each branch must be associated with a `majorVersion`, which is used to filter the git tags when resolving the latest version. When a major version is defined, bumps will fail if the new version crosses the major version boundary.

A separate GitHub workflow is created for each release branch.

Misc: change projen log icon to ✨.

Fixes #734

BREAKING CHANGE: The `releaseWorkflows` is now a map instead of an array.
* **release:** The `initialVersion` option has been removed. The default is `v0.1.0` or `vMM.0.0`.