fix: unit tests side effects, and failures if gpg signing enabled (#854)
Running the unit tests in `bump.test.ts` locally for projen development can cause (funny-looking) errors if you happen to use gpg signing for git. In addition, running the unit tests is affecting global git configuration which has side effects on local development setups (I noticed this when creating the first commit for this pull request...).

This shouldn't affect how our automated builds/tests run on GitHub actions, but it is a QOL improvement for development.