feat(deps): deduplicate dependencies (#862)
Some components are created multiple times (e.g. `UpgradeDependencies` in a default node project is created once to generate a `upgrade-dependencies` task and once more to generate an `upgrade-projen` task). If a component specifies a dependency, the dependency may get added to `deps.json` multiple times. This is not desirable because it's possible for versions/metadata to conflict, and the dependency will get listed multiple times when reading from `project.deps.all`.

This PR changes the default behavior of `project.deps.addDependency()` to override an existing dependency if a dependency has been defined with the same name AND same dependency type.

(I wasn't sure if there are use cases for having the same dep listed under multiple dependency types, but adding de-deplication so that a given dep can only be listed under a single dependency type would be an easy change.)

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.