fix(typescript): cannot collocate sources and compiled output (#868)
Right now, if you want to collocate sources and compiled output, i.e use the same value for `srcdir` and `libdir`:

```typescript
new TypeScriptProject({
  name: 'test',
  defaultReleaseBranch: 'test',
  libdir: 'lib',
  srcdir: 'lib',
  });
```

If fails to build because the tsconfig both includes and excludes the `lib` directory, essentially making it so that there are no sources to compile. 

```json
  "include": [
    "lib/**/*.ts",
  ],
  "exclude": [
    "node_modules",
    "lib"
  ],
```

This PR enables providing the same value and collocating these files.
This makes it easier to onboard projen into existing projects that do have collocation and rely on it (regrettably) 

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.