fix(web): build tasks do not run projen (#867)
Fixes #754

The default behavior for web projects has been updated to run `npx projen` (as well as any compile and test tasks, if those have been defined) before generating a production build via `react-scripts build` or `next build`.

In addition, `project.addTask` has been modified so that it's not possible to override an existing task with the same name. If you want to override an existing task, you can remove it first with `project.removeTask` and then add your task.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.