chore(deps): consolidate upgrade workflow if projen secret is available (#870)
Currently we create two separate upgrade workflows, one for projen and one for everything else. 

The reason is that projen upgrades often require changes to github workflows, which requires special permissions from the token used to create the PR. 

However, if such a token is defined, and it usually is to make projen auto upgrade as well, we can just use that token and merge the two workflows into one. 

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.