feat(projen new): --git and --no-git (without interaction)
Remove the interactive behavior of the CLI around git when running `projen new`. The default new behavior will automatically run `git init` (which is idempotent) and also create an initial commit in the repo with the contents immediately after bootstrapping. If `--no-git` is set, no git operations will be performed.