feat: non-text sample files (#840)
Closes #768.

Introduces new fields "sourcePath" and "sourceDir" for `SampleFile` and `SampleDir` respectively that allows you to specify the path of a file or directory respectively that should be copied from. Through this option, it is now possible to include non-text files in the sample code of a new project.

Also, updates the sample files that are provided in new react projects.