docs: add ARCHITECTURE.md (#903)
Closes #647.

Not a comprehensive doc - just focused on file synthesis for v1. I've tried to keep this at a high level while including details that I thought would be most useful, and hopefully won't be outdated 6 months from now. As always, open to feedback. :-)

Even though the sub-headers are questions, the doc is spiritually meant to be just for architecture things, not a FAQ - open to restructuring things / renaming headers to make this more clear if needed.

I've also tried to avoid linking to source code for the most part in case things get refactored / to avoid stale links.

[View rendered version](https://github.com/Chriscbr/projen/blob/docs/architecture/ARCHITECTURE.md)

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.