feat(release): introduce `release` task
Extract some steps from the GitHub release workflow into a projen task so it can be executed manually via `projen release`.

The `release` task now performs the following: `bump` + `build` + `unbump` + anti-tamper. 

The task will produce two temporary files that can be used to publish a release: `.bump.json.tmp` and `.changelog.md.tmp` (both are .gitignored). The GitHub workflow will pick up those files and create a GitHub release with them.

Next steps: this is an initial step to address #782 and #726. The next step is to add tasks that can be used to publish a release. These tasks can be composed of `publish:github`, `publish:npm`, `publish:nuget`, which we can then use in the GitHub workflow. Additionally, we will likely want to allow manual creation of git tags and committing of a CHANGELOG file when projects are not hosted on GitHub.