feat(release): introduce "release" task (#917)
Extract some steps from the GitHub release workflow into a projen task so it can be executed manually via `projen release`.

The `release` task **prepares a release** by performing the following: `bump` + `build` + `unbump` + anti-tamper. 

The task will emit two files under the `dist/` directory which can be used to publish a release: `version.txt` and `changelog.md` (both are .gitignored). The GitHub workflow will pick up those files (and the rest of the contents under `dist`) and publish them as GitHub releases and to the various package managers.

Next steps: this is an initial step to address #782 and #726. The next step is to add tasks that can be used to publish a release. These tasks can be composed of `publish:github`, `publish:npm`, `publish:nuget`, which we can then use in the GitHub workflow. Additionally, we will likely want to allow manual creation of git tags and committing of a CHANGELOG file when projects are not hosted on GitHub.

Misc: add more information to task inspection and execution outputs (namely, include environment variables).