chore: remove duplicate unit test (#925)
This test doesn't seem to do what it claims since the --no-git flag isn't being passed in, and the description seems identical to the test immediately preceding it. It was probably added by accident.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.