fix(release): typescript release is broken due to missing files
The `package` task for TypeScript deletes the `dist` directory before it runs. This conflicts with the release task that was introduced in #917 which emitted `version.txt` and `changelog.md` into `dist` for later consumption when the GitHub release is created.

To address this, we moved to cleanup of `dist` from the `package` to the `release` task.