fix(release): use posix path for changelog and bump file (#944)
When using projen on a windows machine on a project with the version component, the workflows and tasks contained `\` in their path, thus resulting in a difference between windows/not-windows synths. It also breaks the release script when running in github actions.

This PR forces the use of posix slashes for those files.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.