fix(docgen): typedoc could not find sources (#945)
I'm not sure that this was working for anyone else, but typedoc requires an entrypoint of some sort to be passed in. It also seemed reasonable to update the version of typedoc (the issue was present in both versions) so that it can support typescript 4.3.

As an opinioned change, I added the `disableSources` flag as well. Without that flag, typedoc generates links to the source in the html with a SHA that will be constantly changing. While a nice feature, it causes very noisy diffs.