fix(inventory): Recursive JSII module Lookup  (#937)
When developing external modules we encountered some issues with the projen new command:
1) https://github.com/projen/projen/issues/894 Prevented us to upgrade
2) https://github.com/projen/projen/issues/605 Was a small anoyance that prevented us to have single project packages to work smoothly

Both problems occur with the following package structure:

FeaturePackage => GenericOrganisationProjenPackage => projen

This would end up in the following node_modules structure:
```
  node_modules
  - FeaturePackage
    - node_modules
      -  GenericOrganisationProjenPackage
  - projen
```

Before this PR the logic of the inventory was to scan the full folder of node modules for .jsii modules. It would list all project types and had a exclusive logic to filter out projen native projects. Projen would only see the top level packages with this logic, eventhough the jsii references were correct

This PR changes the call of the discovery function for newProjectFromModule. It will now only pass the requested module. Recusively it will scan for jsii types based on the jsii dependencies. This way it will discover all 3 packages. This will solve issue 1 (and potentially scan a bit faster because it will not search through all package folders now)

As a side optimalisation, it will only list the projects of the requested module to see during the new phase. This means it will allow single project feature packages without project specification.

fixes: 894
fixes: 605

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.